+++ src/lagradii.py
@@ -1,16 +1,23 @@
 """
 Author: A. Lassen
-Last modification: August 1 2025
+Last modification: September 26th, 2025
 
 ------------------- Change log
 14/07/25: Fixed incorrect calculation of kpc/pixel parameters
 15/07/25: SSD calculation has been correctly implemented into class
 24/07/25: Updated methods for inclination correction
 29/07/25: Modularization of calc_lagradii() method
-          Added fill_zeros() utility function for error handling for edge cases in angular bin processing
+          Added fill_zeros() utility function for error handling of edge cases in angular bin processing
 01/08/25: Updated method to calculate SSD, now including handling of uncertainties
+26/09/25: Code version 0.3.1 --> 0.4. Summary of the alterations:
+          - Added support for error arrays (err_dict)
+          - Implemented bootstrap_cumsum() for uncertainty estimation of Lagrangian radii
+          - Refactored _process_abin() to use cumulative interpolation to derive Lagrangian radii
+          - calc_lagradii() now returns uncertainties (r_err, rphys_err) along with radii
+          - Deprecated fill_zeros() function
+          - Improved validation process of inputs
 """
-__version__ = "0.3.1"
+__version__ = "0.4"
 
 import numpy as np
 import pandas as pd
@@ -25,17 +32,59 @@ from astropy.cosmology import FlatLambdaCDM
 cosmo = FlatLambdaCDM(H0=70, Om0=0.3)
 # ----
 
-def fill_zeros(indices):
-    indices = indices.copy()  # avoid modifying the original array
-    for i in range(1, len(indices)):
-        if indices[i] == 0:
-            indices[i] = indices[i - 1]
-    return indices
+# ===================================== //
+def bootstrap_cumsum(tab, percentiles, xcol_name="dr", ycol_name="values",
+                     nboot=2000, sig_y=None, random_state=None, mode="normal"):
+    ### Convert percentiles to fractionals, and set up random generator
+    ps = 0.01 * percentiles
+    rng = np.random.default_rng(random_state)
+
+    ### Unpack x,y arrays from input table columns
+    assert np.all(np.isin([xcol_name, ycol_name], tab.columns)), \
+           f"Columns {xcol_name} and/or {ycol_name} are not included in input table"
+    x = tab[xcol_name].to_numpy()
+    y = tab[ycol_name].to_numpy()
+
+    ### If an y-error is not provided, assume it is sqrt(y)
+    if sig_y is None:
+        sig_y = np.sqrt(np.clip(y, 0, None))
+
+    ### Bootstrapping on cumulative sum
+    samples = {q: [] for q in [f"q{i+1}" for i in range(ps.size)]}
+    for j in range(nboot):
+        if mode == "poisson": pert = rng.poisson(np.clip(y, 0, None))
+        else:
+            pert = rng.normal(y, np.maximum(1e-12, sig_y))
+            pert = np.clip(pert, 0., None) # ensure non-negative values
+
+        tot = pert.sum()
+        if tot == 0:
+            for key in samples.keys(): samples[key].append(np.nan)
+            continue
+
+        cum = np.cumsum(pert) / tot
+        for index, q in enumerate(ps):
+            if q < cum[0] or q > cum[-1]: samples[f"q{index+1}"].append(np.nan)
+            else:
+                samples[f"q{index+1}"].append(np.interp(q, cum, x))
+    
+    ### Return chain with all disturbed arrays and results
+    chains = {}
+    for key in samples.keys():
+        arr = np.asarray(samples[key])
+        arr = arr[~np.isnan(arr)]
+
+        if arr.size == 0: chains[key] = (np.nan, np.nan, arr)
+        else:
+            chains[key] = (np.mean(arr), np.std(arr, ddof=1), arr)
+
+    return chains
+# ===================================== //
 
 class AngSlicing():
     __version__ = __version__
 
-    def __init__(self, imdata, center, mask=None, **kwargs):
+    def __init__(self, imdata, center, mask=None, err=None, **kwargs):
         ### Handle keyword arguments and default options
         self.__set_params__(**kwargs)
 
@@ -43,6 +92,12 @@ class AngSlicing():
         self.xc, self.yc = center
         self.imdata = imdata.copy()
 
+        self.err = err
+        if self.err is not None:
+            assert np.logical_and(self.err.ndim == self.imdata.ndim,
+                                  self.err.shape == self.imdata.shape), \
+                                  "Error array must be identical to input image array!"
+        
         ### Validate bin width
         self.nbin = 360. / self.bin_width
         assert self.nbin % 1 == 0.0, "The angle bin widht must be a multiple of 360."
@@ -74,6 +129,9 @@ class AngSlicing():
         self.imdata[self.imdata < self.min_value] = np.nan
         self.imdata[self.mask] = np.nan
 
+        if self.err is not None:
+            self.err[np.isnan(self.imdata)] = np.nan
+
         return None
     
     def slicing(self):
@@ -90,9 +148,11 @@ class AngSlicing():
         ### Initialize table
         t = {"x": xarray, "y": yarray, "dx": dx, "dy": dy,
              "theta": angles, "values": []}
+        if self.err is not None: t["values_err"] = list()
         
         for iy, ix in zip(yarray, xarray):
             t["values"].append(self.imdata[iy, ix])
+            if self.err is not None: t["values_err"].append(self.err[iy, ix])
 
         tab = pd.DataFrame(t)
         abins = np.linspace(0, 360, self.nbin + 1)
@@ -103,40 +163,38 @@ class AngSlicing():
 class Lag_radii():
     __version__ = __version__
     
-    def __init__(self, im_dict, center, mask_dict, z=None, quiet=False, **kwargs):
+    def __init__(self, im_dict, center, mask_dict, err_dict=None, z=None, quiet=False, **kwargs):
         ### Handle keyword arguments and default options
         self.__set_params__(**kwargs)
         self.quiet = quiet
 
+        if err_dict is not None: self.has_err=True
+        else: self.has_err=False
+
         ### Unpack central coordinates
         self.xc, self.yc = center
-
+        
         ### If redshift is provided, calculate physical scale (kpc/arcsec) at this given distance
         self.z = z
         if self.z is not None:
-            ps_kpc = cosmo.kpc_proper_per_arcmin(self.z).value/60.
+            ps_kpc = cosmo.kpc_proper_per_arcmin(self.z).to("kpc/arcsec").value
             self.kpc_per_pixel = self.px_scale * ps_kpc # [kpc/pixel]
         else:
             self.kpc_per_pixel = 0.0
 
-        ### Validate required keys in both dictionaries
-        required_keys = ["young", "int"]
-        for dict_name, d in [("im_dict", im_dict), ("mask_dict", mask_dict)]:
-            missing = np.array(required_keys)[~np.isin(required_keys, list(d.keys()))]
-            if len(missing) > 0:
-                raise ValueError(f"{dict_name} missing required keys: {missing.tolist()}")
+        ### Validate required keys in all dictionaries
+        self.__validate_dicts__(im_dict, mask_dict, err_dict)
 
-        ### Add images and masks as attributes of the class
+        ### Add images and masks as attributes of the class and validate them afterwards
         self.imy, self.imi = im_dict["young"], im_dict["int"]
         self.m1, self.m2 = mask_dict["young"], mask_dict["int"]
 
-        assert self.imy.ndim == self.imi.ndim == self.m1.ndim == self.m2.ndim == 2,\
-            "Input arrays are not images! (i.e. dim != 2)"
-        assert self.imy.shape == self.imi.shape == self.m1.shape == self.m2.shape,\
-            "Shape of input arrays do not match!"
+        self.__validate_arrays__(err_dict)
 
         # Ensure mask is applied to input images
         self.imy[self.m1], self.imi[self.m2] = np.nan, np.nan
+        if self.has_err:
+            self.erry[self.m1], self.erri[self.m2] = np.nan, np.nan
 
         ### Get initial tables
         self.taby, self.tabi = self._make_tab()
@@ -162,14 +220,52 @@ class Lag_radii():
         for key, value in defaults.items():
             setattr(self, key, value)
 
+    def __validate_dicts__(self, im_dict, mask_dict, err_dict, required_keys=["young", "int"]):
+        for dict_name, d in [("im_dict", im_dict),
+                             ("mask_dict", mask_dict)]:
+            
+            missing = np.array(required_keys)[~np.isin(required_keys, list(d.keys()))]
+            if len(missing) > 0:
+                raise ValueError(f"{dict_name} missing required keys: {missing.tolist()}")
+
+        if self.has_err:
+            missing = np.array(required_keys)[~np.isin(required_keys, list(err_dict.keys()))]
+            if len(missing) > 0:
+                raise ValueError(f"err_dict missing required keys: {missing.tolist()}")
+            
+        return None
+    
+    def __validate_arrays__(self, err_dict):
+        shapes = [self.imy.shape, self.imi.shape, self.m1.shape, self.m2.shape]
+        dim_array = np.array([self.imy.ndim, self.imi.ndim, self.m1.ndim, self.m2.ndim])
+        
+        if self.has_err:
+            self.erry, self.erri = err_dict["young"], err_dict["int"]
+            shapes.extend([self.erry.shape, self.erri.shape])
+            dim_array = np.append(dim_array, [self.erry.ndim, self.erri.ndim])
+
+        assert np.all(dim_array == 2), "Input arrays are not images! (i.e. dim != 2)"
+
+        for i, shape in enumerate(shapes[1:], 1):
+            assert shape == shapes[0], f"Shape mismatch: Array {i} has shape {shape}, expected {shapes[0]}"
+        
+        return None
+
     def _make_tab(self):
         ### Start applying angular slicing in each image
         aslice = {}
-        aslice["young"] = AngSlicing(self.imy, (self.xc, self.yc), mask=self.m1,
-                                     bin_width=self.bin_width, min_value=self.min_value)
-        aslice["int"] = AngSlicing(self.imi, (self.xc, self.yc), mask=self.m2,
-                                   bin_width=self.bin_width, min_value=self.min_value)
         
+        if self.has_err:
+            aslice["young"] = AngSlicing(self.imy, (self.xc, self.yc), mask=self.m1, err=self.erry,
+                                         bin_width=self.bin_width, min_value=self.min_value)
+            aslice["int"] = AngSlicing(self.imi, (self.xc, self.yc), mask=self.m2, err=self.erri,
+                                       bin_width=self.bin_width, min_value=self.min_value)
+        else:
+            aslice["young"] = AngSlicing(self.imy, (self.xc, self.yc), mask=self.m1,
+                                         bin_width=self.bin_width, min_value=self.min_value)
+            aslice["int"] = AngSlicing(self.imi, (self.xc, self.yc), mask=self.m2,
+                                       bin_width=self.bin_width, min_value=self.min_value)
+
         ### Inherit attributes
         self.nbin, self.angles = aslice["young"].nbin, aslice["young"].angles
 
@@ -189,35 +285,7 @@ class Lag_radii():
 
         return taby,tabi
 
-    def calc_lagradii(self, tab):
-        ### Start by sorting table by angle value        
-        tab = tab.sort_values(by="theta", ascending=True)
-        tab.reset_index(drop=True, inplace=True)
-
-        ### Create the outputs
-        shape = ((self.nbin, self.q.size))
-        xslice = np.zeros(shape, dtype=np.float32)
-        yslice, rslice, rslice_phys = np.zeros_like(xslice), np.zeros_like(xslice), np.zeros_like(xslice)
-
-        ### Iterate over angle bins
-        angle_bins = np.arange(0, self.nbin)
-        if self.quiet:
-            for i in angle_bins:
-                xslice[i, :], yslice[i, :], rslice[i, :], rslice_phys[i, :] = self._process_abin(i, tab)
-
-        else:
-            for i in tqdm(angle_bins, desc="Iterating over angle bins", total=self.nbin):
-                xslice[i, :], yslice[i, :], rslice[i, :], rslice_phys[i, :] = self._process_abin(i, tab)        
-        
-        return {"x": xslice,
-                "y": yslice,
-                "r": rslice,
-                "rphys": rslice_phys}
-    
     def _process_abin(self, i, tab):
-        ### Start by initalizing the outputs
-        xslice, yslice, rslice, rslice_phys = [], [], [], []
-
         ### Get sub-table for a given angular bin
         tab_i = tab.loc[tab["bin"] == i].copy()
         tab_i.dropna(inplace=True)
@@ -230,20 +298,8 @@ class Lag_radii():
         if self.rlim is None: tot_flux = tab_i["values"].sum()
         else: tot_flux = tab_i["values"].loc[tab_i["dr"] <= self.rlim].sum()
 
-        ### Iterate over the rows of sorted sub-table
-        ### Obs.: Quartiles expressed in terms of flux percentages
-        qflux = self.q * (tot_flux/100.)
-
-        f,k = 0,0
-        indices = np.zeros(self.q.size, dtype=np.int32)
-        for j, row in tab_i.iterrows():
-            f += row["values"]
-            if f >= qflux[k]:
-                indices[k] = j
-                k += 1
-            
-            if k > self.q.size - 1: break
-        if tab_i.shape[0] <= indices.size: indices = fill_zeros(indices)
+        tab_i["cum_values"] = tab_i["values"].cumsum()     ### Cumulative sum
+        tab_i["cum_frac"] = tab_i["cum_values"] / tot_flux ### Cumulative fraction
 
         ### Get list of tables, sliced at different Lag. radii                
         theta1 = i * self.bin_width
@@ -251,24 +307,68 @@ class Lag_radii():
         mean_theta_rad = 0.5*(theta1 + theta2)
 
         if tab_i.empty:
-            xslice.append(0.)
-            yslice.append(0.)
-            rslice.append(0.)
-            rslice_phys.append(0.)
-
+            return (np.repeat(0., self.q.size), np.repeat(0., self.q.size),
+                    np.repeat(0., self.q.size), np.repeat(0., self.q.size), None, None)
         else:
-            rslic_tables = [tab_i.iloc[np.arange(0, r+1)] for r in indices]
-            for rslic_tab in rslic_tables:
-                rslic = rslic_tab["dr"].max()
+            cumfrac_values = tab_i["cum_frac"].clip(0., 1.) # Ensure there are no rounding problems
+            rslice_i = np.interp(0.01 * self.q, cumfrac_values, tab_i["dr"])
+            
+            err = tab_i["values_err"].to_numpy() if self.has_err else None            
+            rslice_chains = bootstrap_cumsum(tab_i, self.q, sig_y=err)
 
-                rslice.append(rslic)
-                xslice.append(rslic * (-1.) * np.sin(np.deg2rad(mean_theta_rad)))
-                yslice.append(rslic * (-1.) * np.cos(np.deg2rad(mean_theta_rad)))
+            xslice_i = (-1.) * rslice_i * np.sin(np.deg2rad(mean_theta_rad))
+            yslice_i = (-1.) * rslice_i * np.cos(np.deg2rad(mean_theta_rad))
 
-                if np.isin("dr_kpc", list(tab.columns)): rslice_phys.append(rslic_tab["dr_kpc"].max())
-                else: rslice_phys.append(0.)
+            if np.isin("dr_kpc", list(tab.columns)):
+                rslice_phys_i = self.kpc_per_pixel * rslice_i
+                
+                rslice_phys_chains = {}
+                for key in rslice_chains.keys():
+                    rslice_phys_chains[key] = (self.kpc_per_pixel * rslice_chains[key][0],
+                                               self.kpc_per_pixel * rslice_chains[key][1],
+                                               self.kpc_per_pixel * rslice_chains[key][2])
+            
+            else: rslice_phys_i = np.zeros_like(rslice_i)
 
-        return xslice, yslice, rslice, rslice_phys
+            return (xslice_i, yslice_i,
+                    rslice_i, rslice_phys_i,
+                    rslice_chains, rslice_phys_chains)
+    
+    def calc_lagradii(self, tab):
+        ### Start by sorting table by angle value
+        tab = tab.sort_values(by="theta", ascending=True)
+        tab.reset_index(drop=True, inplace=True)
+
+        ### Create the outputs
+        shape = ((self.nbin, self.q.size))
+        xslice = np.zeros(shape, dtype=np.float32)
+        yslice, rslice, rslice_phys = np.zeros_like(xslice), np.zeros_like(xslice), np.zeros_like(xslice)
+        rslice_err, rslice_phys_err = np.zeros_like(rslice), np.zeros_like(rslice_phys)
+
+        ### Iterate over angle bins
+        angle_bins = np.arange(0, self.nbin)
+        if self.quiet:
+            for i in angle_bins:
+                abin_results = self._process_abin(i, tab)
+                xslice[i, :], yslice[i, :], rslice[i, :], rslice_phys[i, :], chains_r, chains_rphys = abin_results
+                
+                if chains_r is not None:
+                    rslice_err[i, :] = np.array([chains_r[key][1] for key in chains_r.keys()])
+                    rslice_phys_err[i, :] = np.array([chains_rphys[key][1] for key in chains_rphys.keys()])
+                else: pass
+        else:
+            for i in tqdm(angle_bins, desc="Iterating over angle bins", total=self.nbin):
+                abin_results = self._process_abin(i, tab)
+                xslice[i, :], yslice[i, :], rslice[i, :], rslice_phys[i, :], chains_r, chains_rphys = abin_results
+
+                if chains_r is not None:
+                    rslice_err[i, :] = np.array([chains_r[key][1] for key in chains_r.keys()])
+                    rslice_phys_err[i, :] = np.array([chains_rphys[key][1] for key in chains_rphys.keys()])
+                else: pass
+
+        return {"x": xslice, "y": yslice,
+                "r": rslice, "rphys": rslice_phys,
+                "r_err": rslice_err, "rphys_err": rslice_phys_err}
 
     def rot_and_stretch(self, x, y, PA, inc, ang_type="deg"):
         if ang_type == "deg":
@@ -313,33 +413,6 @@ class Lag_radii():
         islice_icorr["rphys"] = self.kpc_per_pixel *  np.sqrt(islice_icorr["x"]**2. + islice_icorr["y"]**2.)
 
         return yslice_icorr, islice_icorr
-
-    # def deprojection(self, PA, inc, ang_type="deg"):
-    #     if ang_type == "deg":
-    #         PA = np.radians(PA)
-    #         inc = np.radians(inc)
-
-    #     xcone_int, ycone_int, rcone_int, rphys_int = self.int_cones
-    #     xcone_young, ycone_young, rcone_young, rphys_young = self.young_cones
-
-    #     ### Initialize outputs
-    #     xcy_icorr, xci_icorr = np.zeros_like(xcone_young), np.zeros_like(xcone_int)
-    #     ycy_icorr, yci_icorr = np.zeros_like(ycone_young), np.zeros_like(ycone_int)
-    #     rcy_icorr, rci_icorr = np.zeros_like(rcone_young), np.zeros_like(rcone_int)
-
-    #     ### Apply deprojection by rotating and stretching x,y coordinates of the cones
-    #     for i in range(self.q.size):
-    #         xcy_icorr[:, i], ycy_icorr[:, i], rcy_icorr[:, i] = \
-    #             self.rot_and_stretch(xcone_young[:,i], ycone_young[:,i],PA, inc, ang_type="radians")
-    #         xci_icorr[:, i], yci_icorr[:, i], rci_icorr[:, i] = \
-    #             self.rot_and_stretch(xcone_int[:,i], ycone_int[:,i], PA, inc, ang_type="radians")
-
-    #     ### Also convert deprojected rcones to physical units
-    #     rcy_phys_icorr = np.sqrt((self.kpc_per_pixel * xcy_icorr)**2. + ((self.kpc_per_pixel * ycy_icorr)**2.))
-    #     rci_phys_icorr = np.sqrt((self.kpc_per_pixel * xci_icorr)**2. + ((self.kpc_per_pixel * yci_icorr)**2.))
-
-    #     return [(xcy_icorr, ycy_icorr, rcy_icorr, rcy_phys_icorr),
-    #             (xci_icorr, yci_icorr, rci_icorr, rci_phys_icorr)]
 
     def calculate_SSD(self, yslice, islice, re_arcsec, start=1, errors=False):
         ### Attribute unit to Re